pyproject.toml
[tool.poetry]
name = "anamnesisai"
version = "0.1.0"  # semantic-release
description = "AI-driven anamnesis collection system in the healthcare domain."
readme = "docs/index.md"
authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>", "Satarupa Deb <satarupa2212@gmail.com>"]
packages = [
  {include = "anamnesisai", from="src"},
]
license = "BSD 3 Clause"
exclude = [
  ".git/*",
  ".env*",
]

[tool.poetry.dependencies]
python = ">=3.9,<4"
openai = ">=1.52.2"
fhir-resources = {version = ">7",  allow-prereleases = true }
torch = [
  {version = ">=2.5.0", source="pytorch-cpu"},
]
torchvision = [
  {version = ">=0.20.0", source="pytorch-cpu"},
]
rago = {version = ">=0.9.0", extras=["cpu"]}
eval_type_backport = {version = ">=0.2", python = "<3.10"}
typeguard = ">=4"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.2"
pytest-cov = ">=4.1.0"
coverage = ">=7.2.7"
pre-commit = ">=3.3.2"
ruff = ">=0.2.0"
mypy = ">=1.6"
bandit = ">=1.7.8"
vulture = ">=2.7"
mccabe = ">=0.6.1"
ipython = "<8"
ipykernel = ">=6.0.0"
Jinja2 = ">=3.1.2"
mkdocs = ">=1.4.3"
mkdocs-exclude = ">=1.0.2"
mkdocs-jupyter = ">=0.24.1"
mkdocs-literate-nav = ">=0.6.0"
mkdocs-macros-plugin = ">=0.7.0,<1"
mkdocs-material = ">=9.1.15"
mkdocstrings = ">=0.21.2"
mkdocstrings-python = ">= 1.1.2"
mkdocs-gen-files = ">=0.5.0"
makim = "1.19.0"
python-dotenv = ">=1.0"


[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"


[tool.pytest.ini_options]
testpaths = [
    "tests/",
]
markers = [
    "skip_on_ci: Skip these tests in CI",
]

[tool.ruff]
line-length = 79
force-exclude = true
src = ["./"]
exclude = [
  'docs',
]
fix = true

[tool.ruff.lint]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "D",   # pydocstyle
  "YTT", # flake8-2020
  "RUF", # Ruff-specific rules
  "I001",  # isort
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
# Use a single line between direct and from import
lines-between-types = 1

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
exclude = ["scripts/"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
  "langchain_experimental.sql",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
